# backend/openapi.yaml
openapi: '3.0.3'
info:
  title: flux - OpenAPI-Document
  description:
    This is the OpenAPI 3.0 specification for the flux backend-API.
  version: v0
servers:
  - url: 'http://localhost:8620/'
...
paths:
...
  /api/v0/index/records:
    get:
      summary: list available records
      description:
        Returns a list of available records for the given query.
      tags:
        - index
      security:
        - sessionCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/requestId'
        - name: search
          in: query
          description: search filter
          required: false
          schema:
            type: string
        - name: type
          in: query
          description: content type
          required: false
          schema:
            $ref: '#/components/schemas/RecordType'
        - name: range
          in: query
          description: request subset of records
          required: false
          schema:
            type: string
            pattern: '/[0-9]+-[0-9]+/'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/ResponseMd'
                  content:
                    type: object
                    properties:
                      count:
                        type: integer
                        description: number of records matching filters
                      records:
                        type: array
                        description: array of records
                        items:
                          $ref: '#/components/schemas/Record'
                    required:
                      - count
                      - records
                required:
                  - meta
  /api/v0/index/record/{recordId}:
    get:
      summary: get record info
      description: Returns record information.
      tags:
        - index
      security:
        - sessionCookieAuth: []
      parameters:
        - name: recordId
          in: path
          description: record identifier
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/requestId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/ResponseMd'
                  content:
                    allOf:
                      - $ref: '#/components/schemas/Record'
                      - type: object
                        properties:
                          content:
                            oneOf:
                              - $ref: '#/components/schemas/VideoInfo'
                              - $ref: '#/components/schemas/SeriesInfo'
                              - $ref: '#/components/schemas/CollectionInfo'
                required:
                  - meta
  /api/v0/index/record/{videoId}:
    get:
      summary: get record info
      description: Returns record information.
      tags:
        - index
      security:
        - sessionCookieAuth: []
      parameters:
        - name: videoId
          in: path
          description: video identifier
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/requestId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/ResponseMd'
                  content:
                    allOf:
                      - $ref: '#/components/schemas/Record'
                      - type: object
                        properties:
                          content:
                            oneOf:
                              - $ref: '#/components/schemas/VideoInfo'
                              - $ref: '#/components/schemas/SeriesInfo'
                              - $ref: '#/components/schemas/CollectionInfo'
                required:
                  - meta
  /api/v0/index/video/{videoId}:
    get:
      summary: get video info
      description: Returns video information.
      tags:
        - index
      security:
        - sessionCookieAuth: []
      parameters:
        - name: videoId
          in: path
          description: video identifier
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/requestId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/ResponseMd'
                  content:
                    $ref: '#/components/schemas/VideoInfo'
                required:
                  - meta
  /api/v0/index/record/{recordId}/current-video:
    get:
      summary: get info on current video
      description:
        Returns information on current video for a given record.
      tags:
        - index
      security:
        - sessionCookieAuth: []
      parameters:
        - name: recordId
          in: path
          description: record identifier
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/requestId'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/ResponseMd'
                  content:
                    type: object
                    properties:
                      video:
                        $ref: '#/components/schemas/VideoInfo'
                      playback:
                        type: object
                        description: additional playback information
                        properties:
                          timestamp:
                            type: number
                            description: playback progress in seconds
                            minimum: 0
                            example: 85
                required:
                  - meta
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: fluxSession
  parameters:
    requestId:
      name: requestId
      in: query
      description: request identifier
      required: false
      schema:
        type: string
  schemas:
    RequestId:
      type: string
      description: request identifier
    ResponseMd:
      type: object
      description: response metadata
      properties:
        id:
          $ref: '#/components/schemas/RequestId'
        ok:
          type: boolean
          description:
            whether request has been processed successfully
        error:
          type: object
          description: problem description in case of error
          properties:
            code:
              type: integer
              description: error-code
              example: 400
            short:
              type: string
              description: short error description
              example: Bad request
            long:
              type: string
              description: long error description
              example: Missing required argument in request body.
          required:
            - code
            - short
            - long
      required:
        - ok
    Record:
      type: object
      description: record information
      properties:
        id:
          type: string
          description: record identifier
        type:
          $ref: '#/components/schemas/RecordType'
        name:
          type: string
          description: record name
        description:
          type: string
          description: record description
        thumbnailId:
          type: string
          description: thumbnail identifier
      required:
        - id
        - type
    RecordType:
      type: string
      description: record type identifier
      enum:
        - series
        - movie
        - collection
    VideoInfo:
      type: object
      description: video information
      properties:
        id:
          type: string
          description: video identifier
        name:
          type: string
          description: video name
        description:
          type: string
          description: video description
        metadata:
          type: object
          description: video/codec metadata
          additionalProperties: true
        thumbnailId:
          type: string
          description: video thumbnail identifier
        trackId:
          type: string
          description: track identifier
    SeriesInfo:
      type: object
      description: series details
      properties:
        seasons:
          type: array
          description: array of seasons for this series
          items:
            type: object
            description: seasion information
            properties:
              id:
                type: string
                description: season identifier
              name:
                type: string
                description: season name
              episodes:
                type: array
                description: array of seasons in this season
                items:
                  $ref: '#/components/schemas/VideoInfo'
        specials:
          type: array
          description: array of specials for this series
          items:
            $ref: '#/components/schemas/VideoInfo'
    CollectionInfo:
      type: array
      description: array of videos for this collection
      items:
        $ref: '#/components/schemas/VideoInfo'
...